requirements.txt
streamlit==1.20.0
pandas==1.5.3
gspread==5.4.0
google-auth==2.18.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
openai==0.27.0
langchain==0.0.118
langchain-experimental==0.0.19
pandasai==0.2.0
requests==2.28.2
langchain_groq
